move server only libraries into $lib/server Fixes #12
