+++ starter/src/hooks.server.ts
@@ -1,4 +1,4 @@
-import { auth } from '$lib/lucia/mysql';
+import { auth } from '$lib/server/lucia/mysql';
 
 export const handle = async ({ event, resolve }) => {
 	event.locals.auth = auth.handleRequest(event);
