+++ starter/src/lib/server/drizzle/mysql/models/tokens.ts
@@ -1,5 +1,5 @@
-import { drizzleClient } from '$lib/drizzle/turso/client';
-import { emailVerification, passwordResetToken } from '$lib/drizzle/turso/schema';
+import { drizzleClient } from '$lib/server/drizzle/mysql/client';
+import { emailVerification, passwordResetToken } from '$lib/server/drizzle/mysql/schema';
 import { eq } from 'drizzle-orm';
 import { generateRandomString, isWithinExpiration } from 'lucia/utils';
 
@@ -68,11 +68,9 @@ const generatePasswordResetToken = async (userId: string) => {
 };
 
 const validateEmailVerificationToken = async (token: string) => {
-	const storedToken = await drizzleClient
-		.select()
-		.from(emailVerification)
-		.where(eq(emailVerification.id, token))
-		.get();
+	const storedToken = (
+		await drizzleClient.select().from(emailVerification).where(eq(emailVerification.id, token))
+	)[0];
 
 	if (!storedToken) {
 		throw new Error('Invalid token');
@@ -93,11 +91,9 @@ const validateEmailVerificationToken = async (token: string) => {
 };
 
 const validatePasswordResetToken = async (token: string) => {
-	const storedToken = await drizzleClient
-		.select()
-		.from(passwordResetToken)
-		.where(eq(passwordResetToken.id, token))
-		.get();
+	const storedToken = (
+		await drizzleClient.select().from(passwordResetToken).where(eq(passwordResetToken.id, token))
+	)[0];
 
 	if (!storedToken) {
 		throw new Error('Invalid token');
