+++ starter/src/lib/server/emails/send.ts
@@ -1,6 +1,6 @@
 import { dev } from '$app/environment';
-import { sendTestEmail } from '$lib/emails/mailhog';
-import { sendResendEmail } from '$lib/emails/resend';
+import { sendTestEmail } from '$lib/server/emails/mailhog';
+import { sendResendEmail } from '$lib/server/emails/resend';
 
 const sendEmail = async (options: { from: string; to: string; subject: string; html: string }) => {
 	if (!dev) {
