+++ starter/src/routes/(auth)/password-reset/+page.server.ts
@@ -1,7 +1,7 @@
 import { TRANSACTIONAL_EMAILS_ADDRESS, TRANSACTIONAL_EMAILS_SENDER } from '$env/static/private';
-import { generatePasswordResetToken } from '$lib/drizzle/mysql/models/tokens';
-import { getUserByEmail, getUserProfileData } from '$lib/drizzle/mysql/models/users';
-import { sendEmail } from '$lib/emails/send';
+import { generatePasswordResetToken } from '$lib/server/drizzle/mysql/models/tokens';
+import { getUserByEmail, getUserProfileData } from '$lib/server/drizzle/mysql/models/users';
+import { sendEmail } from '$lib/server/emails/send';
 import { getFeedbackObjects } from '$lib/utils';
 import { fail } from '@sveltejs/kit';
 import { z } from 'zod';
