+++ starter/src/routes/api/oauth/github/+server.ts
@@ -1,5 +1,5 @@
 import { dev } from '$app/environment';
-import { githubAuth } from '$lib/lucia/mysql';
+import { githubAuth } from '$lib/server/lucia/mysql';
 import { redirect } from '@sveltejs/kit';
 
 export const GET = async ({ cookies }) => {
